Adicionar seção 12.3 (#46)
* Adicionar seção 12.3

* Add comments for section 12.1

Adding comments for existing section 12.1 to improve reading.

* Add part of section 12.3 code

Adding initial elements of the section, as well as the
cost funcions described.

* Complete section 12.3

Adding costs and greedy algorithm in basic version and
version with cost added

* Attempt to prove termination for parts

Add theorem trying to prove termination por
parts in 12.1, but the prove was not completed.

* Update README.md

Complete sections 12.1 and 12.3

* Fix typo